fix: update unique index in PatientDeviceIdentifier since PatientSerNum was replaced by Username
Revises the PatientDeviceIdentifier table to replace the unique index on `['PatientSerNum', 'DeviceId']` with one on `['Username', 'DeviceId']`. This is necessary since the PatientSerNum field was deprecated in favor of Username, and since the listener uses this index, having a special behavior for inserting into this table with `ON DUPLICATE KEY UPDATE`.

* Note: since unique indexes are strictly enforced in the DB, I had to add a sanitization query that deletes all old rows that would violate the new index. This is required to prevent the revision from failing if any problematic rows exist.

See merge request opalmedapps/db-docker#67